Move GraphQL mutations in a separate class
